Finishes the @owl-factory web worker module
