+++ src/@owl-factory/web-worker/index.ts
@@ -1,0 +1,2 @@
+
+export { newWebWorker } from "./utilities";
